src/requirements.txt
Django==1.11
psycopg2==2.6.2
redis==2.10.5
coreapi==2.3.3
pygments==2.2.0
markdown==2.6.10
django-extensions==1.7.6
django-enumfields==0.8.2
djangorestframework==3.6.3
drf-extensions==0.3.1
transmissionrpc==0.11
celery==4.0.2
gunicorn==19.7.1
coverage==4.4.1
coveralls==1.2.0
ipython==5.3.0
ipdb==0.10.3
